r3430 another patch by Jeroen Benda:
In database.diff you will find an update of the listCells function to
change the sorting of the racks in the tab Rackspace for an object.
Sorting was row_id and is now alphabetical on row_name. This file also
contains an update where the properties for an object are not sorted by
type first, but by name so you can group properties together which are
dictionary and string.