r4893 Address the foreign key issue with AttributeValue table (no bug ID, issue discussed on the mailing list). MySQL documentation explains:
"The handling of foreign key references to nonunique keys or keys that
contain NULL values is not well defined for operations such as UPDATE
or DELETE CASCADE. You are advised to use foreign keys that reference
only UNIQUE and NOT NULL keys."

In particular, AttributeValue.attr_id referencing AttributeMap.attr_id
was wrong in this regard (rev 4674, release 0.19.7). Referencing the
unique (type_id, attr_id) tuple is the right solution, although this
requires a copy of object type ID in the AttributeValue table. This
commit implements the required changes.