initial support for Cisco UCS
Signed-off-by: Denis Ovsienko <infrastation@yandex.ru>

This commit is based on Jeff's patch he contributed to RackTables as a
GPL piece of work. I had to justify most of fine details in the code
and its layering, but the overall approach was preserved.

UCS is presently the only use case of queryDevice() with HTTP-supplied
username and password possible. This commit is an initial attempt to
handle such use cases.

New code:

* getObjectContentsList(): returns a list of object's nested objects
* ucsTranslatePushQueue(): maps PHP pseudocodes to UCS command(s)
* ucsReadInventory(): converts a plaintext listing of components into a
  PHP array
* renderEditUCSForm(): renders the HTML contents of the new "UCS" tab
* autoPopulateUCS(): ophandler to import the contents of a UCS domain
* cleanupUCS(): ophandler to remove the contents of a UCS domain
* trigger_ucs(): a trigger enabling the "UCS" tab on UCS domains' object
  page
* gateways/ucssdk: a Python backend to UCS SDK

Updated code:

* interface-lib.php: register a new PNG image
* dictionary.php: introduce a server chassis pseudo-model for a UCS
  domain
* install.php: update the default value of IPV4OBJ_LISTSRC option
* resetUIConfig(): update resetting of IPV4OBJ_LISTSRC option for
  consistency
* navigation.php: register new conditional "UCS" tab of an object's page
  and its associated functions
* detectDeviceBreed(): update to detect a new "ucs" breed
* $breedfunc: register the functions of the new "ucs" breed
* queryTerminal(): update for the new "ucs" breed and its first "ucssdk"
  protocol

Further work is required to align the RackTables data model with regard
to management instances and their association with physical objects, but
this is so far left out of scope of this implementation.