rename one of two renderFilePreview() funcs (#603)
There used to be two function with the same name, one in solutions.php
and another in interface.php. Since the module dispatcher in index.php
treats these two files mutually exclusive, this name clash remained
unnoticed until a user forced unconditional inclusion of interface.php
in a plugin file. This commit renames the function in solutions.php to
renderImagePreview().