allow setting MYSQL_ATTR_MAX_BUFFER_SIZE (#659)
There are several bottlenecks getting in the way of large files properly
uploaded and downloaded, MySQL client buffer size being one of them
(others are MySQL server buffer size and assorted PHP limits).

This change makes it possible to tune the application in this regard
without changing the source code.

* init_config(): add a commented pdo_bufsize option to secret.php
* connectDB(): do custom setup with PDO constructor driver options