speed-up displaying 802.1Q queue summary
The problem with this summary is that it requires to perform
considerConfiguredConstraint for each switch. On large installations
this leads to a poor rendering time of main 802.1Q page. The solution
is to cache the disabled sync state in last_errno DB field.
This changeset turns the 'disabled' 802.1Q state to a real state, and
makes syncdomain.php to operate with it.

syncdomain.php: set the disabled state for those switches which do not
comply to SYNC_802Q_LISTSRC. Force sync of those switches which are in
disabled state, but do not compy to SYNC_802Q_LISTSRC.

detectVLANSwitchQueue: E_8021Q_SYNC_DISABLED leads to 'disbaled' state
get8021QDeployQueues: change the output format, do not group by
	enabled/disabled key
render8021QStatus: do not use enabled/disabled keys
renderDeployQueue: idem