bugfix: remote ifname shortening in CDP/LLDP
Since we need to know the device breed of the remote device to shorten
its ifname, it is hard to be done in deviceconfig handlers. So the
getlldpstatus/getcdpstatus queries now return the original remote port
names. The shortening is performed in renderDiscoveredNeighbors, where
the remote object id is already known.