Fix install.php usage of DELIMITER
DELIMITER apparently cannot be used in a Pdo context and was giving an error during the installation on all recent MySQL versions. I fixed the install.php to work more like upgrade.php with the SQL statements saved in an array. This also makes it possible to easily re-use future DML/DDL changes from upgrade.php in install.php