bugfix: 802.1Q support of Huawei VRP 8.80 devices
vrp85Read8021QConfig was pasing output of the "display vlan summary"
command taking the last line by regexp "VLAN ID: .*". Huawei VRP 8 has 3
classes of VLANs in that output: static, dynamic, reserved. RT wants to
read only static list, but it was expecting the rest two classes to be
empty.

The recent version of VRP 8.80 fills the third class (it is called
'service' VLANs) with VIDs 4064-4094, and this breaks the parser.

This fix forces RT to take the first list, not the last one.