readChapter(): make PHP sorting case-insensitive
Since the commit 4aa8609 the function had no ORDER BY in the SQL query
it uses to retrieve rows from the dictionary chapter because the wiki
link parsing has to be performed before the sorting. asort() was
put in place to do the job, however, its processing is case-sensitive by
default.

This change makes asort() do case-sensitive sorting for consistency with
other functions that deliver rows from the database in the exact order
of ORDER BY.