don't allow to turn objects into locations (#1339)
Locations, rows and racks are special in that normal object management
should work around them even though they are represented in Object
table. This was properly implemented when adding new objects but object
properties editor allowed to set object type to a Location-specific
type, which was a bug.

This change factors the existing object type filter out to a helper
function and makes both forms use it.

* withoutLocationTypes(): new function, the helper
* renderAddMultipleObjectsForm(): use the helper
* getObjectTypeChangeOptions(): ditto