UI: links to rows are displayed on 'manage' tab
It is hard to enter a row's page if you have hundreds of racks:
you need to wait until the entire rackspace is rendered to select the
row.

There is a compact list of rows on the 'Manage rows' tab, but no links
to rows there. This change adds links to a rows into rendered table.

- $SQLSchema has changed: the 'rackc' field added to the 'row' realm.
- getAllRows: deprecated, superseded by listCells ('row')
- renderRackspace, renderEditRackForm: don't use getAllRows
- renderRackspaceRowEditor: prints link to a row.