retrieve Munin graphs based on FQDN too (#1279)
In Munin objects are always grouped by domain and the domain is what
follows the hostname (host.do.ma.in). To retrieve a Munin graph it is
necessary to know both plus the name of the graph. Before this change
the object's displayed name was considered for this purpose, now FQDN
will override it when set.

* getMuninNameAndDomain(): new function instead of plain preg_split()
* proxyMuninRequest(): update to use the above
* renderObjectMuninGraphs(): ditto, and report some errors with an error
  message, not the exception screen