new feature: vlan domain groups
This is useful when you want to isolate some VLANs to
a subset of a network topology, but have switches that
should see all of them. The example topology is:

D -- C -- E
    / \
   A   B
  /\   /\
 ...   ...

Here we create two domains: domain A for the A switch and all of its
downsteam peers, and the domain B, respectively.

We want VLANs from both domains A and B to propagate to the switches D and E
through the switch C.
We want to prevent a VLAN from the domain A to be set up or propagated over the
switch B. There should be some VLANs that exists in both A and B, thus
propagating everywhere.

It is not convenient to set up thes constaints through permissions ACL and 802.1Q
template VLAN masks. This kind of configurations would be not visual and tend to
become unsyncronized.

Instead, we create the domain group "AB" and set up the domains A, B as its members.
The switches that should see all the VLANs from both domains (C,D,E) whould have an
order in the domain AB, while the others - in one of the (A, B) domains.

These changes were made to the code:
new field 'group_id' added to the VLANDomain table
getVLANDomainStats: group domain's statistics include the members',
	new 'subdomc' result field.
getVLANDomain: new 'subdomc' result field.
getDomainGroupMembers: new DB func
getDomainVLANs: merges the VLAN lists of members into the group master results
getDomainVLANList: idem
getVlanRow: new result field 'domain_group_id'
sameDomains: boolean function to check a domain border crossing
checkPortRole: use sameDomains
renderVLANDomainListEditor: domain groups editor added
renderVLANDomain: for a group master, displays all the vlans of group members
renderVLANDomainVLANList: display a domain's own vlans only
the opspec 8021q-vdlist-upd removed, updateVLANDomain ophandler added