mind the references in patch cables tables
The patch cables configuration used to rely on foreign key exceptions
only to mind existing relations to connectors and cables. Other parts of
the interface typically in addition to that disable the removal link for
items depended by other tables. This change makes the "Edit connectors"
and "Edit cable types" tabs consistent with that behaviour.

* getPatchCableConnectorList(): extend the SQL query with a sub-SELECT
  to yield the reference counter
* getPatchCableTypeList(): ditto
* renderSimpleTableWithOriginEditor(): handle the "refc" key when it is
  present