printTagCheckboxTable(): restore TD COLSPAN
The sequence of TRs representing the tag tree with checkboxes used to
have the TDs with COLSPAN=2, several functions relied on this to keep
their headers and footers in line. Commit cef98fd lost the COLSPAN
attribute and made the tables look a little bit off. Now this is back
to normal.