add new functions for HTML escaping
The long-existing niftyString() function works OK for TABLE TDs but does
not work well when it comes to input elements because those cases
actually require different processing. The best solution seems to be to
have multiple functions, one for each purpose, and to have all other
code eventually migrated from niftyString() to those.

* stringForLabel()
* stringForTD()
* stringForTextInputValue()
* stringForTextarea()