Added 'comment' field to getObjectAttrsSearchResults
This allows for searching the "comments" field of objects (VMs etc).  This doesn't return the comment field as part of the search results; I think this is actually OK, as comment fields can be *huge.*

RackTables search allows searching by comment in many types of resources, but not VMs, other objects. This brings objects in line with the rest.  Also submitted as a patch at https://bugs.racktables.org/view.php?id=1489