Allow the ucs gateway to import rackmount servers
Newer versions of the Cisco UCS system can manage rackmount servers that
are linked with a Cisco VIC network card to the Fabric Interconnects.
By adding a new element (ComputeRackUnit) to gateways/ucssdk it is able
to detect rackmount servers by Cisco (C-Series).

In ophandlers.php the function autoPopulateUCS() is extented to create
the according server object.