generalize the graph cycles resolver
The tag forest editor tab used to have a separate form for resolving
graph cycle issues in the TagTree database table. This change reworks
the form into a standalone generic tab that makes it possible for the
user to fix similar data structures given these are declared in the
function.

* getOrphanedTags(): generalize and rename to getInvalidNodes()
* renderTagTreeEditor(): handle the forest structure only
* renderGraphCycleResolver(): new function with the tabhandler
* renderDataIntegrityReport(): amend
* triggerGraphCycleResolver(): new function