queryTerminal: 'terminal echo-mode line' for vrp55
Huawei devices that apply to the 'vrp55' breed echoes all the input
commands back to terminal in single-byte chunks. This causes too
many TCP segments to be sent by switch if it gets a plenty of verbose
commands.

'terminal echo-mode line' command disables this behavior. It does not
exist in vrp85 software series.