optimize compareDecomposedPortNames
sortPortList() turned out to be a heavy call for large number of ports
to sort. Avoid numerous calls to numCompare/numSign in
compareDecomposedPortNames, just do the final result normalization w/o
subsequent calls. This reduces the cumulative function runtime from 48ms
to 40ms in test environment.

Also, add the optional parameter sorted=True to the
getObjectPortsAndLinks() function. It allows to skip sorting the result
if it is not necessary.