fix an edge case in object highlighting
Within all parts of a rack diagram "state_Thw" is the intended class
name for an object that both has "has_problems" flag raised and is the
current object. As far as I remember at least some versions of
RackTables implemented this properly.

Apparently this logic was broken some time ago as commits a2037cd and
5a7bf29 attempted to put the class name right for zero-U objects and
objects within a chassis respectively. Moreover, the class name still
was wrong for ordinary objects within a rack.

This change re-implements the convention.

* highlightObject(): fix the logic for ordinary objects in a rack by
  prepending the class suffix instead of replacing
* printObjectDetailsForRenderRack(): fix the logic for objects in a
  chassis