enable object highlighting in rack thumb images
This change extends the existing rack thumb image functions to allow
an optional "object_id" parameter to make the object's pixels on the
output image brighter (just like it is already implemented in rack
diagrams). Using this feature automatically delivers the image through
the "midirack" venue, which uses no caching.

The new parameter is not used by RackTables code at the moment but it
should be useful for plug-ins.

* getRackThumbLink(): accept an extra argument
* dispatchImageRequest(): detect and supply an extra argument
* printRackThumbImage(): accept an extra argument