tests: refine GetChildrenListTest
Change the class to keep the complete list of the IDs being added on
setup and to delete the rows one by one in the reverse order on
teardown. This way the teardown method is more compact and the foreign
key constraints remain satisfied all the time.

Use commitDeleteObject() to delete the objects and locations as this
covers any associated rows from related tables. In particular, before
this change ObjectHistory would contain stray rows after each test run.