rectify error handling in UI config
The functions in database.php shouldn't imply they were called in an
ophandler context as it can be any other context, hence change them to
throw InvalidArgException instead of InvalidRequestArgException and
change respective calling functions in ophandlers.php to use newIRAE()
to present the input validation errors with the same exception class as
before.

Also in database.php do not specify default value for arguments, use
array_fetch() and a helper variable to simplify the code, reword some
error messages and do not involve mb_strlen() in an empty string test.

* database.php
  * setConfigVar(): also dismiss unused 3rd argument
  * setUserConfigVar()
  * resetUserConfigVar(): also dismiss useless 'is_userdefined' check
* ophandlers.php
  * updateUI()
  * saveMyPreferences()
  * resetMyPreference()
  * saveQuickLinks()