add type guards to array_values_same()
Unless both arguments to array_diff() are arrays, the function returns
an empty array as if the arguments were arrays and the first one was a
subset of the second. In this case (though there would be a PHP
warning) array_values_same() would return TRUE to indicate no
differences between the arguments.

Add a few tests.