factor serializeTagStats() out
This change fixes an issue raised on the racktables-users mailing list
where the user would want to delete a tag but wouldn't be able to see
why exactly the editor does not allow to do this. Now both the default
and the editor tabs of the tag tree (forest) page are able to show the
same hint with this information.

* serializeTagStats(): factor out of renderTagRowForViewer()
* renderTagRowForViewer(): update
* renderTagRowForEditor(): update