conform to MySQL NO_ZERO_DATE SQL mode, pt. 1
MySQL versions prior to 5.7 typically have NO_ZERO_DATE SQL mode
disabled, whilst version 5.7 typically has it enabled. RackTables
versions up to and including 0.20.11 contain a few SQL columns that have
the default value set to "0000-00-00 00:00:00", hence those versions
will fail to install on a MySQL server with NO_ZERO_DATE enabled.

Interestingly, the same RackTables versions will continue to work if
installed on a MySQL server without NO_ZERO_DATE and later changed to
with it (likely as an implicit effect of MySQL upgrade from pre-5.7 to
5.7).

One possible solution to this problem could be to start managing the
server SQL mode during DB connection setup. But as it is difficult to
predict which way MySQL will evolve in this regard, a better long-term
solution seems to be just not to use the zero date default value (and
any other mode-dependent language features in general) at all.

This commit changes the first of the two involved tables -- LDAPCache,
in which the column last_retry is used only as input to TIMESTAMPDIFF(),
which returns NULL when the argument is NULL or the zero date. This way
the change should not be visible in the associated PHP code (though I
have not tested it).

* install.php
* upgrade.php