Travis CI: use SQL queries from the installer
The SQL dumps in tests/data for each respective release contain the
initial schema and minimal data plus the contents of
init-sample-racks.sql on top. This way using even the most recent dump
from that directory would (and in fact did) fail the builds made after
changes to install.php. For example, a pull request that makes changes
to the schema would likely fail the test as the database would belong to
a release version.

This commit changes the test setup script to initialize the database
the same way as the installer would do, and to load the current
init-sample-racks.sql after that. This allows to test a part of the
actual path the installer takes and resolves the failed pull requests
issue.