Travis CI: don't load the sample data
The sample data may interfere with the outcome of the tests in either
way (false negative or false positive) and it would not reproduce
consistently between the Travis CI temporary copy and the developer's
working copy because of non-test data being different.

To make this more deterministic it seems better to initialize the
Travis CI database without the sample data and to expect all normal
dependencies of a test to be managed by the test class methods. This way
it should be easier to correlate the cause and the effect.