Travis CI: do load the sample data, but later
Loading the sample data dump into a freshly generated RackTables
database is actually a useful test by itself so it wasn't right to drop
it in the previous commit. As the data can interfere with the tests,
these just should be run before loading the dump, not instead
(especially as the unit tests are assumed to tear all their fixtures
down as they complete). And as soon as the sample data dump cannot be
easily unloaded, it does not belong to the express test set, which is
purposed for repeated runs in the same working copy by a developer.

In other words, it is a valid test that belongs only to a specific step
of the Travis CI script.