renderTableViewer(): support more customization
Add an optional argument to override the default set of attributes of
the TABLE element when required. Make it possible to specify custom
class for each individual TR element through the "_tr_class" array
element. This proved to be useful in a local plug-in and can
potentially be reused in the main code.