Travis CI: the folding did not work, use a script
For some reason folding did not work as expected in the "before_script"
section. So long as I could not find a way to make it work afterwards,
move the relevant shell commands to a separate script of its own, which
now also allows to have a better script with shell constructs.