refine some age formatting code
The function formatAge() used to derive the amount of seconds from
time() and to format that as text. Move the main body of the code
involved to a new function, formatAgeSeconds(), which accepts the
amount of seconds as a parameter, such that the function is now pure.
Fixup the code to return "min" for minutes in all cases. Leave the old
function name as an alias for formatAgeTimestamp(), which calls time()
and implements the old semantics. Add some tests.