getImageHREF(): fix bugs after commit d5f8fa1
PHP Notice:  Undefined index: error in interface-lib.php on line 419
PHP Notice:  Undefined index: width in interface-lib.php on line 432
PHP Notice:  Undefined index: height in interface-lib.php on line 433

When called with an invalid $tag argument value, the function would
yield the PHP messages and return HTML code with a malformed SRC
attribute value in the IMG element. The image would subsequently fail
to render because the HTTP response for the SRC URI request would be
not an image but an HTML page with error message.

Rewrite the code to prepare input to makeHtmlTag() rather than to
produce any HTML directly (a side effect is that ampersands in the URI
are now properly encoded). For undefined image tags specify an URI that
returns a standard error image, add an undefined index guard.