optimize write in callScript()
If a huge buffer was passed into callScript (tens of Mb), it took much
time to write it into pipe using small 4k blocks, even if system not
blocking us. That was because of re-allocating this huge string buffer ($in variable)
on every write iteration ($in = substr($in, $written)).

Now changed to tracking write cursor in the constant $in buffer and
making a small substring (4k) on every iteration.