Travis CI: allow PHP 7.0 to fail
This is because it seems to be broken on Travis CI side, as the same
commit that passed the build a few days ago (012f09a) does not pass
it (as well as subsequent commits) with an error that has to do with
PHPUnit class autoloading:

PHP Fatal error:  Class 'PHPUnit_Framework_TestCase' not found in
/home/travis/build/infrastation/racktables/tests/ConfigVarTest.php
on line 3

In my working copy (PHP 7.0.15) the tests pass as before. This change
does not exclude PHP 7.0 from the Travis CI build matrix so it should
be easy to see the recovery later.

While at it, reduce the git clone depth to reduce the amount of
unnecessary work.