replace port triggers with table locking
It turned out after the 0.20.12 release that error handling in MySQL
triggers works differently across different installations. Several users
had reported that trying to call a non-existent stored procedure from a
trigger may fail in more than one way. The call often fails during the
trigger run time when the L2 address voilates the custom constraint (and
thus rolls the transaction back as originally intended) but sometimes it
fails at the trigger creation time (and thus fails to create the
trigger). On my installation it seems to fail on UPDATE even when the
l2address column is not involved.

Another issue was, the transaction in syncObjectPorts() used SELECT ...
FOR UPDATE to lock the rows that is was going to update or delete, but it
was not clear if the constraint would be reliably enforced for all newly
inserted L2 addresses. This way two or more concurrent transactions could
potentially execute their triggers successfully independently of each
other and insert conflicting rows afterwards.

This change addresses the above issues by switching from triggers back to
table-level locking with PHP constraint implementation. It preserves the
recently added ability to make multiple changes within a single table
lock through addDesiredPort() and syncObjectPorts(). As before, the whole
set of changes does not happen if any individual change fails a sanity
check.

* getObjectPortsAndLinksTerse(): add a new helper function to avoid table
  locking complexity
* commitAddPort(): Restore the use of alreadyUsedL2Address() and the
  locking removed in commit e6b7635.
* commitUpdatePort(): idem
* convertPDOException(): remove the trigger-specific error code
* alreadyUsedL2Address(): reinstate the revision before commit e6b7635
* exceptions.php: remove the now unused L2AddressException class
* syncObjectPorts(): replace the transaction with a table lock and
  rearrange the code to have all checks done before making any changes
* upgrade.php: drop the triggers
* install.php: ditto