implement ISO 27001 risk valuation
This commit implements risk valuation -- the first step in the process
defined in the ISO 27001 information security management system. The
valuation uses criteria, criterion groups, per-object values and
criticality to produce a single resulting score for each RackTables
object (or rack) designated as an ISO 27001 asset.

This data is stored in a set of 8 new SQL tables, which is mostly
self-contained. The only direct relation to the rest of the schema is
through one object ID foreign key, which also works for racks as they
are in the same table.
* ISO27001AssetGroup
* ISO27001AssetOwner
* ISO27001AssetMaintainer
* ISO27001Asset
* ISO27001CriterionGroup
* ISO27001Criterion
* ISO27001CriterionGroupValueSet
* ISO27001AssetCriterionValue

This feature introduces 6 new configuration variables. 11 new autotags
make it possible to distinguish particular criteria, criterion groups
and asset groups/owners/maintainers in RackCode.

The user interface now includes the following new elements:
* an administrator's page under "Configuration",
* a data entry operator's tab on the object (rack) page,
* an itemized valuation table on the default tab of an object (a rack)
  page, and
* two reports on the "Reports" page.

The new tab is off by default and can be enabled with a RackCode-based
option. Most of the new interface code goes into interface-iso27001.php:
* renderISO27001Configuration()
* renderISO27001AssetGroupEditor()
* renderISO27001AssetOwnerEditor()
* renderISO27001AssetMaintainerEditor()
* renderISO27001CGroupEditor()
* renderISO27001CValueSetEditor()
* renderISO27001CriteriaEditor()
* renderObjectISO27001Asset()
* renderObjectISO27001CValues()
* renderISO27001ValuationReport()
* renderISO27001StatisticsReport()

Changes to the database-layer code mostly consist of new functions
that access the new SQL tables:
* getISO27001AssetGroupList()
* getISO27001AssetOwnerList()
* getISO27001AssetMaintainerList()
* getISO27001CriterionGroupList()
* getISO27001CriterionGroupValueSets()
* getISO27001CriterionList()
* getISO27001AssetInfo()
* getISO27001AssetList()
* getObjectISO27001CValues()
* commitObjectISO27001CValues()

The new functions that do formatting and other processing of ISO 27001
data (including the risk valuation) go into functions.php:
* formatISO27001CVAsLabel()
* formatISO27001CriteriaAsLabel()
* getISO27001CriterionGroupOptions()
* getISO27001AssetGroupOptions()
* getISO27001AssetOwnerOptions()
* getISO27001AssetMaintainerOptions()
* getISO27001Configuration()
* getISO27001AssetValuation()

Changes to ophandlers.php are as simple as new $opspec_list declarations
except operations on assets and criteria values, which require dedicated
new functions to implement the new autotags:
* addObjectISO27001Asset()
* delObjectISO27001Asset()
* updObjectISO27001Asset()
* saveObjectISO27001CValues()

Other changes not mentioned elsewhere:
* pi.css: add new styles
* install.php: usual changes
* upgrade.php: usual changes
* code.php
  * checkAutotagName(): amend to include new autotags
* triggers.php
  * triggerISO27001CGroupExists(): new function for the admin page
  * triggerObjectISO27001Asset(): new function for the data entry tabs
  * triggerObjectISO27001CValues(): idem
* interface-reports.php
  * renderDataIntegrityReport(): list new foreign keys
* database.php
  * commitResetObject(): delete ISO 27001 rows to release the FK
  * generateISO27001AssetAutotags(): new function
  * generateEntityAutoTags(): add a call to the above
* interface.php
  * renderRackInfoPortlet(): add rendering of the ISO 27001 elements
  * renderObject(): idem
  * renderSimpleTableEditor(): new generic helper function
  * requestedISO27001Details(): new helper function
  * generateISO27001SummaryLines(): new function to return ISO 27001
    relations and score where applicable
  * renderObjectISO27001Portlet(): new function to itemize the valuation
    score if requested
* ophandlers.php
  * resetUIConfig(): include the new configuration variables