better display objects that have no common name
Address a long-standing glitch: when a port on the currently displayed
object was (or was suggested to be) linked to a port of an object that
had no name, the displayed remote object name would be an empty string.
This change makes the interface display a standard substitute instead,
which among other things makes the remote object hyperlink clickable.

* fetchPortList(): add remote object type ID to the result columns
* renderObjectPortRow(): make use of formatObjectDisplayedName()
* renderPortsForObject(): idem
* findSparePorts(): all of the above