improve CodeMirror integration
Reimplement the RackCode language mode in rackcode.js using simpler
JavaScript code and define a standalone "rackcode" theme in rackcode.css
instead of overriding the default theme. In renderRackCodeEditor() pass
more properties to CodeMirror constructor and make minor HTML fixups.
In renderRackCodeViewer() rewrite the code to use CodeMirror in read-only
mode and to scroll to requested line on request. Amend the URL format in
refRCLineno() respectively.