make the object properties form more user-friendly
updateObjectAttributes() used to feed each new attribute value to MySQL
connection and let it raise an exception for an invalid input. The
exception would be dispatched as a hard error with the MySQL error
codes and a stack trace page.

Improve this by making calls to genericAssertion() as required and
converting any resulting IRAE exceptions to a human-readable form.

Add helper methods getValue() and getReason() to InvalidArgException.