refine displaying of attributes
formatAttributeValue() was initially purposed for objects and when it
yielded a hyperlink it would hard-code "page=depot" into it. When the
function was later reused for racks, rows and locations, this logic
didn't deliver working URLs.

Add necessary checks such that this feature only tries to work when it
can work, add a date format mouse hint, add minimal reporting of RackCode
parsing errors and move some code around for clarity.

Modify the functions below as required:
* renderRow()
* renderRackInfoPortlet()
* renderObject()
* renderSearchResults()
* renderLocationPage()