refine editing of attributes
It used to be possible to map "date" attributes to racks, rows and
locations, but the properties editor would not know how to handle
those attributes in particular and would not make permission checks
correctly in general.

This change addresses those issues by replacing four similar code
blocks with one new function that can handle all types of attributes
for all types of objects and makes necessary permission checks.

* renderEditAttributeTRs(): a new universal function
* renderEditRowForm(): use the new function instead of custom local code
* renderEditObjectForm(): idem
* renderEditRackForm(): idem
* renderEditLocationForm(): idem
* updateObjectAttributes(): update request argument checks to match
  what the current implementation puts into the form