put usort_portlist() right (GH #199)
The input to sortPortList() is an array of arrays, if it is an array of
anything else the function (since commit 07d22ce) will use array_fetch()
on the elements anyway, causing issues downstream. Vladimir Sukhonosov
had reported numerous PHP warnings in this regard.

Fix usort_portlist() to feed proper argument into sortPortList().