refine the SQL column type for tag colour
MySQL char(6) can contain any string including a hex notation of a 24-bit
RGB colour. With this storage type the PHP code should check if the value
is valid every time it reads the column, and if the value is invalid
there is no good way to deal with it as the invalid value could have been
stored a long time ago and it could be caused by a bug in a
non-interactive script so there is nowhere to put the blame.

Based on the prior experience it is better to eliminate this place for an
error earliest possible, do this by switching to an unsigned 24-bit
integer type in the MySQL table, this way any non-NULL value stored in
the column is always valid and any attempt to store an invalid value will
fail at one or another level immediately so there is a feedback loop now.

* get_pseudo_file(): update CREATE TABLE
* upgrade.php: modify ALTER TABLE
* getTagList(): amend SELECT to convert to hex at server side
* commitUpdateTag(): amend column conversion
* $opspec_list: refine declarations for tagtree-edit-createTag
* isHTMLColor(): reduce to a small unary function and add tests
* HTMLColorForDatabase(): new converter function with tests
* HTMLColorFromDatabase(): idem
* assertHTMLColorArg(): simplify and merge into genericAssertion()