improve renderDataIntegrityReport() a little more
In the triggers section add the entries below to the list of known
triggers and compare the list with the items in the schema both ways.
Use renderTableViewer() to display the difference.

* EntityLink-before-insert
* EntityLink-before-update

In the foreign keys section make a similar change and amend the new
plugins to return their lists of FKs in the new format.

* RackThumbnail-FK-rack_id
* VLANDomain-FK-group_id